pyproject.toml
[project]
name = "spendcast-benchmark"
version = "0.1.0"
description = "A chatbot implementation using a local LLM with MCP protocol integration"
authors = [
    {name = "Spendcast Team", email = "team@spendcast.ai"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.13.0",
    "httpx>=0.24.0",
    "pydantic-settings>=2.0.0",
    "rich>=13.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''
